Devops Sync Job: Add WasScheduled field to database of builds (#5133)
For the devops sync job we only want to open bugs for scheduled builds. This is because devs often misconfigure or try weird things with manual builds.

In order to accomplish this we need to add to the builds database a field for if the build was scheduled or not. I played around with the api and the easiest way to do this is to check who the build is requested for. If it is requested by `` then it is a scheduled build.