Longhaul Tests: Add tolerance for unauthorized exceptions (#5157)
We get a few unauthorized exceptions in longhaul which Varun and I believe to be service errors. He has recommended that we add a tolerance so that is what I am doing in this PR.

Also, I took this as an opportunity to fix the unmatchedResultsCount returning a value smaller than the true value. This happened because we were getting the unmatchedResultsCount from the collection of unmatchedResults, but this collection size is capped.