Nested Connectivity Tests: Test custom mqtt topics (#5135)
Tests are missing for custom mqtt topics in connectivity. Adding them in this PR. The tests are failing due to product issues, but the bug autologging will catch this and open a bug for product investigation.

Note: Lots of commits below are not relevant to this diff and show because I stacked diffs to accommodate multiple in-flight PRs.