Update YAML pipelines to 1ES-hosted agents (#5175)
This change ports #4844 and #5006 to release/1.2. See those PRs for details.

Additionally, this change updates the `IoTEdge_check` test to use Process.RunAsync for consistency and simplicity.

Finally, this change restores the (undocumented) logging conventions that have previously been in place for the end-to-end tests. Specifically, the logging output users see in the console (DevOps UI or the command line) should be tidy and concise. To achieve this, only information-level logging is shown in the console by default. Information-level logging is reserved for two purposes:
1. **Assures the user that the tests are actually using the arguments they passed in** (like IoT Hub name, device ID prefix, etc.). Most args are confirmed via messages logged in `Profiler.Run` (see # 2), but there are a couple lone `Log.Information` commands that provide feedback about device creation/deletion.
2. **Gives the user some indication that the test is either progressing along or stuck**. This means that most asynchronous commands in the tests output something when they complete, usually via the `Profiler.Run` method. There's also a single `Log.Information` call to announce that a new test is starting.

Any other logging should probably use `Log.Verbose`. End-to-end runs in DevOps always publish a log file that contains all verbose output; check the run's artifacts. The same can be achieved from the command line by including the `logFile` argument in context.json. If you temporarily want to see verbose output in the console, you can set the `verbose` argument to `true` in context.json, but this should _not_ be merged to the YAML pipeline files upstream.