Fix host cpu metric incorrectly reported at 100% (#5204) (#5220)
We made some changes to the system resource collection, where we called slightly different sysinfo apis in an attempt to optimize. This led to what I am guessing is a bug in the sysinfo crate.

We can quickly resolve it if we roll back the commit that made these sysinfo modifications. However we should also keep the main improvement which was to set the open file descriptor limit.