S360: Fix Tls1.2 errors in deployment files (#5225)
Need to make some fixes to some deployment files to fix their tls1.2 configurations. We should explore making this the default for edgehub, but it is potentially a breaking change which needs to be discussed in more detail. For now we are fixing it in the deployment files.

There are some remaining tls1.2 errors in http endpoints that this change won't fix. Still investigating those.