E2E: Fix failing custom mqtt proxy test (#5156)  (#5229)
The proxy tests are failing because the custom mqtt test test uses the TRC, which needs to call iothub for initializing itself. This causes a problem because the TRC was not configured to use the proxy config.

As a solution I have added the 1 line fix to initialize the TRC container with the proxy config.
https://github.com/Azure/iotedge/pull/5156