Metrics Collector: Fix adoption stats (#5190)
We removed the test specific logic from the metrics collector. In doing so, this removed a get twin call that was relied on for gauging usage stats.

The solution is to recreate the `ModuleClient` (after product info and potentially all other fields are set) even in the Log Analytics code path. This change needs some synchronization as to not interfere with metrics potentially being published from the `IotMessage` metrics upload path (which also uses `ModuleClient`). 

To achieve this synchronization, I chose to create a wrapper around the `ModuleClient` which serves as the shared reference across multiple `Tasks`. Multiple of these `Tasks` might use the `ModuleClient`. One `Task` could be recreating it while the other could be attempting to publish messages. In order to achieve synchronization I chose to use a `SemaphoreSlim` because I needed some async lock.

I ran into some issues when attempting to recreate the `ModuleClient`. 
1. If messages tried to be published too quickly, the initial message send would hang.
2. When calling `CloseAsync()` the SDK threw an exception. (SDK bug imo)

I managed to resolve these issues by not calling `CloseAsync()` and waiting 1 minute to publish messages. I confirmed in `EdgeHub` logs that multiple connections did not exist.