Make sure to dispose HttpContentStream when done reading module logs. (#5241)
This another attempt to find and fix a potential issue where we don't dispose sockets in time.

For `GetModuleLogs` operation, we provide `HttpCompletionOption.ResponseHeadersRead` when calling `HttpClient.SendAsync()`. In this case, the response is not loaded into memory at once, but instead can be streamed. It implies, that we also need to properly dispose the stream so the connection is released ASAP.