Custom MQTT Test Module: Implement exponential backoff for reporting to test orchestrator (#5228) (#5251)
I have implemented exponential backoff for reporting to the TRC (orchestrator). This is needed for a few reasons:
1. Longhaul has transient network issues, even in the internal docker network
2. E2E sometimes runs into cases where the test start delay is not long enough, leading to the mqtt test module failing with no retries before the TRC is up and running.

I have also taken this as an opportunity to fix a component governance issue caused by the hyper dependency in the `trc-client` crate. I have bumped the version of hyper, which has in-turn bumped the tokio version and resolved the issue.