Nested E2E: Re-enable nested priority queue tests (part 2) (#5317)
The code changes to the test framework and the test modules were made in #5303. Now this PR will actually re-enable the test for both the proxy and the nested cases. I expected there to be another issue with nested cases, but it seems the tests pass. Not sure why it was disabled in the first place.