Devops Service: Auto assign bugs to owners (#5334)
We have bug generation tooling for failing tests. The next step here is to assign them to the responsible party. This can be done by assigning a bug to the commit author of the failing build. If commits are batched, the assignee will be responsible for determining if any potential failures are caused by their changes and triaging if needed.

In order to do this, the existing bug creation abstraction (`BugManagement`) has been modified via composition to contain two more abstractions. Specifically these steps were taken:
1. Parse the commit from from the devops api response and store in the VstsBuild datatype object.
2. Use new abstraction `CommitManagement` that can take a commit and return a full name. This full name is determined through the git config of the committer at commit time. Since microsoft open source guidelines specify to declare your full name on your account, this should be a safe assumption. 
3. Use new abstraction `UserManagement` to take the committer full name and look up the email address for this microsoft identity.
4. Once we have the email address of the committer we can open a bug on them.

Here is an example bug:
https://msazure.visualstudio.com/One/_workitems/edit/10570189

Also I took this as an opportunity to add isa95 to the devops service.