Custom Mqtt Tests: Better support multi-child/single-parent scenario (#5362)
Nested connectivity tests have multiple child (L3) nodes sharing the same parent (L4). The way the custom mqtt tests are currently structured, messages are intended to flow from the child, to the parent, then back to that same child. But each of these child edge devices publishes/subscribes on this on this same topic for the parent (L4) node. So essentially, a message will emanate from L5 to L4 as intended. But when it comes back, it doesn't just come back to the L5 it was sent from. It gets sent to all child devices.

This design works, but the downside is that child edge devices will receive the telemetry for all other child edge devices running at that time. Then we have logic to filter by a batch id so that we don't receive these messages from other message-pathways.

It would be an improvement if these child edge devices did not share the same topic for the parent (L4) node. Then the L5 test module logs would not be cluttered with irrelevant messages coming from other L5 devices. 

In this PR I have made the following changes:
1. Instead of taking in as input an `initiate_topic`, we now take `topic_suffix`. I hardcoded `initiate` into the topic as I expect this will never change.
2. If a `topic_suffix` is not provided, then we use the randomly generated batch id as part of the topic. This is what will allow the nested connectivity multi-child/single-parent scenario to not share topics on the L4 node.
3.  For the `Relay` case, we now relay everything that starts with `initiate` keyword. So if a topic is published on `initiate/123` it would get relayed back down as `relay/123`.

I also updated the readme which could help to give more context:
https://github.com/Azure/iotedge/pull/5362/files#diff-10d935bdce3ca405feb2e51580e75da0522b4023af05dbcd35109884919a2ff8R1