fix 1.1.5 (#5442)
docker API on windows is checking if socket exists during container creation. Since sockets are created when starting a container (hence after create) the modules are never deployed.

We suspect the issue was not found during testing because of the order of the tests (no proof but it is a very likely scenario).
1. First the old edgeAgent image was used for testing. In this case the create container would work because the legacy workload socket is created from the start. The in the sub-sequent create, the module socket would get created.
2. Then I would test with the new image, since the previous test did create the sockets, create would not complain.

Test:
Same test on windows 2019 and linux.
Verified windows 10 boots.

Tested with the new image from the beginning and verified that sockets are created.
Tested with legacy image and verified that sockets are created and are not used.
Tried stopping a module and checked listener would be stopped.
Tried removing a module with and without stopping the module first.