Use 1ES-hosted agent for CentOS-7 in end-to-end tests (#5285)
1ES-hosted pools now support CentOS-7 agents, so this change migrates the CentOS job in our end-to-end test pipeline to use a 1ES-hosted agent.

I also had to update some test code that discovers the local operating system to determine how to install IoT Edge (e.g., which package manager to use). The test was shelling out to `lsb_release`, which is not guaranteed to be installed. With this change, the tests use the environment variables defined in `/etc/os-release` instead. The release/1.1 branch got this update some time back, but the other branches weren't updated.

I confirmed that the job runs (and passes) on the expected agent in the end-to-end test pipeline.