Use 1ES hosted agent for amd64 single-node connectivty tests (#5456)
Migrate the single node connectivity test pipeline from self-hosted devops agents to 1ES-hosted agents.

Also, update the edgelet Cargo.lock to reference an aziot-identity-service commit with a PACKAGE_VERSION in packages.yaml, that matches the version of aziot-identity-service that the edgelet package depends on. This fixes an error in the connectivity tests, where aziot-edge fails to install.