Patch a vulnerability in System.Text.Encodings.Web (#5484)
One of our dependencies, prometheus-net.AspNetCore, has a transitive dependency on a vulnerable version of System.Text.Encodings.Web. Updating prometheus-net.AspNetCore wouldn't change anything, so the recommended approach is to make an explicit reference to the patched version of System.Text.Encodings.Web, to override the transitive version.

I used `dotnet list package --include-transitive` before and after the change to verify that the transitive dependency is replaced by the explicit dependency within the project.