Add RocksDB_MaxManifestFileSize env var to Edge Agent and Edge Hub.
The RocksDB MANIFEST file stores the history of operations. When it reaches
the configured max size, RocksDB rolls over to a new file and marks
the previous file as obsolete, so that it will be deleted in
the next compaction.

Since the default max size in RocksDB is 1 GiB which might be too high for
some users, this env var allows those users to set it lower.