Enable ARM32 E2E Test After Mitigating Docker Network Issue (#5561)
All Failures in ARM32 were occuring due to failure of a Module to be installed due to the container being previously attached to the aziot docker network. This is a known bug in docker https://github.com/docker/cli/issues/1891.

The problem is mitigated by restarting docker service on  uninstall of edge daemon during teardown sequence.