Make tests resilient to Windows service manager errors (#5608)
In a failed end-to-end test run on Windows earlier today, I noticed that all test failures looked like this:

```
  Error Message:
   System.InvalidOperationException : Cannot stop iotedge service on computer '.'.
  ----> System.ComponentModel.Win32Exception : The service cannot accept control messages at this time.
  Stack Trace:
     at System.ServiceProcess.ServiceController.Stop()
   at Microsoft.Azure.Devices.Edge.Test.Common.Windows.EdgeDaemon.InternalStopAsync(CancellationToken token) in D:\a\_work\1\s\test\Microsoft.Azure.Devices.Edge.Test.Common\windows\EdgeDaemon.cs:line 107
   at Microsoft.Azure.Devices.Edge.Test.Common.Windows.EdgeDaemon.ConfigureAsync(Func`2 config, CancellationToken token, Boolean restart) in D:\a\_work\1\s\test\Microsoft.Azure.Devices.Edge.Test.Common\windows\EdgeDaemon.cs:line 65
   ...
```

I've noticed this before as well. Sometimes, the service manager in Windows isn't ready when you stop a service, but waiting a little while and trying again tends to solve it (e.g., see [this](https://docs.microsoft.com/en-us/previous-versions/ms833805(v=msdn.10)?redirectedfrom=MSDN)). I noticed we weren't retrying when we get this exception, so this change adds the retry logic.

The error is not very common, so I was unable to confirm that my change works around it. But I ran the Windows jobs in the pipeline 7 times (with several jobs running in parallel too), and the stop logic still works, so at least generally I know I didn't make it worse. I added a verbose log so that we can gather more data if we still see this error in the future.