Add RocksDB_MaxManifestFileSize env var to Edge Agent and Edge Hub. (#5610)
Cherry-pick from master of 2c878635c75d3eaddd55c06ae5fd9229d9fe463b and
2385a9b9c6dafd2f72f474e4c742083ed27b3ad3

The RocksDB MANIFEST file stores the history of operations. When it reaches
the configured max size, RocksDB rolls over to a new file and marks
the previous file as obsolete, so that it will be deleted in
the next compaction.

Since the default max size in RocksDB is 1 GiB which might be too high for
some users, this env var allows those users to set it lower.