Upgrade to Azure.Messaging.EventHubs library in end-to-end tests (#5622)
We currently use the Microsoft.Azure.EventHubs library in our CI, end-to-end, connectivity, and long haul tests to read D2C events that land in IoT Hub. That library has been deprecated, however, in favor of Azure.Messaging.EventHubs.

I upgraded the library in our release/1.1 end-to-end tests while doing some experimentation. Since it integrated cleanly and seemed to work just as well, I'd like to commit the changes. We can use this as a pattern for upgrading all other uses in the repo as well, across all branches.

I ran my changes on a limited set of end-to-end tests locally, many times. I also ran the 1.1 end-to-end test pipeline, which was green on all platforms except one test failed on Windows for an unrelated reason.