E2E: Don't error on component governance for proxy test suite (#5694)
On release 1.2 we re-enabled the proxy test suite (#5648) and everything is working fine. When this change was ported to master, the tests fail due to component governance. 

This wasn't caught in the tests because the test for my PR branch didn't use component governance. My current understanding is that it only gets auto-injected for branches in the iotedge repo (and thus not PR branches)
https://github.com/Azure/iotedge/pull/5648#issuecomment-942759541

The reason this works in release/1.2 is because we don't error on component governance. For master this was changed within the last month (#5572).

Waiting on this test run before merge:
https://dev.azure.com/msazure/One/_build/results?buildId=47839241&view=results



# Azure IoT Edge PR checklist: