Fix edgehub_queue_len counting (#5692)
This fixes the count of eh_queue_len by using only the metrics as counters and not setting through other means. We will only increment on propose and decrement on commit. We initialize the count based on how many entries are in the queue from the checkpoint's offset. We still update the count whenever we cleanup the queue.

NOTE
Below graph is with TTL of 7200 seconds and the change in color is when edgeHub was restarted.