Remove WebSocket Ping KeepAlives (#5623) (#5689) (#5723)
The Java clients do not handle the Kestrel WebSocket Ping (KeepAlive) gracefully and can put the client into a bad state. This change sets the interval to Zero which disables the WebSocket Ping (KeepAlive).

This change should be safe because underlying AMQP and MQTT protocols have already built-in Keep Alive mechanisms. Also, IoT Hub does not support/send WebSocket Ping as well, that's why the issue is reproducible only with Edge Hub.

The plan will be to correctly support this in the Java clients going forward but the best remediation that will have the biggest impact is to turn these off and have the edge act similarly to the hub.

Cherry-pick #5689