Update dependency on vulnerable package (#5750)
We recently upgraded a dependency in one of our projects to Microsoft.Extensions.Logging.Console/5.0.0. That package has a transitive dependency on a vulnerable version of System.Text.Encodings.Web. There is no newer version of Microsoft.Extensions.Logging.Console yet, so the recommended approach is to make an explicit reference to the patched version of System.Text.Encodings.Web, to override the transitive version.

I used `dotnet list package --include-transitive` before and after the change to verify that the transitive dependency is replaced by the explicit dependency within the project.

## Azure IoT Edge PR checklist: