Port parallelization changes from images pipeline to  images-release (#5827)
Port changes previously made in #5547 from images.yaml to images-release.yaml. This is a quick port of the changes in order to support the 1.2.5 release.  A follow-on effort should be made to reduce duplication between build-images.yaml and images-release.yaml.

A pipeline was run in ADO to test these changes and the pipeline succeeded. As another test, the images produced by the test run of the pipeline (i.e. edge-agent, edge-hub, edge-simulated-temperature-sensor, and edge-diagnostics) were deployed manually to a VM edge device and all modules/containers launched and ran without any errors. I also ran an E2E pipeline using the edge-agent and edge-hub images built from this modified images-release pipeline and all tests passed.

## Azure IoT Edge PR checklist:

This checklist is used to make sure that common guidelines for a pull request are followed.

### General Guidelines and Best Practices
- [x] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).
- [x] Title of the pull request is clear and informative.
- [x] Description of the pull request includes a concise summary of the enhancement or bug fix.

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.
- Description of the pull request includes 
	- [x] concise summary of tests added/modified
	- [x] local testing done.  

### Draft PRs
- Open the PR in `Draft` mode if it is:
	- Work in progress or not intended to be merged.
	- Encountering multiple pipeline failures and working on fixes.

_Note: We use the kodiakhq bot to merge PRs once the necessary checks and approvals are in place. When it merges a PR, kodiakhq converts the PR title to the commit title, PR description to the commit description, and squashes all the commits in the PR to a single commit. The net effect is that entire PR becomes a single commit. Please follow the best practices mentioned [here](https://chris.beams.io/posts/git-commit/#:~:text=The%20seven%20rules%20of%20a%20great%20Git%20commit,what%20and%20why%20vs.%20how%20For%20example%3A%20) for the PR title and description_