Cherry-pick: Port parallelization changes from images pipeline to  images-release (#5838)
Port changes previously made in #5547 from images.yaml to images-release.yaml. This is a quick port of the changes in order to support the 1.2.5 release.  A follow-on effort should be made to reduce duplication between build-images.yaml and images-release.yaml.

A pipeline was run in ADO to test these changes and the pipeline succeeded. As another test, the images produced by the test run of the pipeline (i.e. edge-agent, edge-hub, edge-simulated-temperature-sensor, and edge-diagnostics) were deployed manually to a VM edge device and all modules/containers launched and ran without any errors. I also ran an E2E pipeline using the edge-agent and edge-hub images built from this modified images-release pipeline and all tests passed.

## Azure IoT Edge PR checklist: