Nested Tests: Agent lock resiliency (#5836)
This PR changes the PAT secret to be the one used elsewhere in the repository. It also adds an initial check that the script can retrieve agent data from the devops api, so the root cause will be more obvious if it happens again. 

I tested the fail path locally and the success path in the pipeline:
https://dev.azure.com/msazure/One/_build/results?buildId=48992747&view=logs&j=23024f98-a3b7-5786-11c6-69db2f5ae7ab&t=37796a81-43e1-5fe6-128f-b89b6495147b

## Azure IoT Edge PR checklist: